requirements.txt
azure-core==1.30.0
azure-storage-blob==12.19.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==42.0.2
flashtext==2.7
idna==3.6
isodate==0.6.1
joblib==1.3.2
nltk==3.8.1
pycparser==2.21
python-dotenv==1.0.1
regex==2023.12.25
requests==2.31.0
six==1.16.0
tqdm==4.66.1
typing_extensions==4.9.0
urllib3==2.2.0
